moved util functions to util.go
